Added interface with SCIP solver
The Virtual Machine can now generate LP format files that comply with CPLEX format.
Console and server detect whether SCIP is in the PATH environment variable, check whether the executable can be found, and if so, permit connecting to the SCIP solver.
Data exchange appears functional, but has not been tested extensively yet.