Icons for sorted charts
Future version will permit sorting variables in charts. This will not change the series data, but reorder the time step sequence.

When multiple variables are sorted in a chart, each will be plotted on its own transformed time scale.

When sorting a variable "alone", the other variables in chart are plotted on the untransformed time scale.

When sorting a variable "leading", all variables in the chart are plotted on the transformed time scale.

When multiple variables are sorted "leading", the time scale is transformed on the multi-level sort of these variables, where the plotting order determines first basis for comparison, second basis for comparison, etc.