Improved customization (from data) of composite constraints
Bound lines of composite constraints now have their own properties modal. This permits entering a list of sets of point coordinates (as text of as remote data file) which can then be viewed as bound lines "variants" 1, 2, etc.
Per bound line, multiple selectors can be defined, each associated with an expression that will be used at run time to compute the index of the set of point coordinates to be used for this bound line.